feat(chatbot): refactor chatbot initialization and update server configuration
