+++ apps/chatbot-ui/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "mongodb-chatbot-ui": "*",
+    "mongodb-chatbot-ui": "^0.14.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
