+++ apps/chatbot-ui/tsconfig.json
@@ -5,6 +5,5 @@
     "baseUrl": ".",
   },
   "include": ["src"],
-  "exclude": ["node_modules", "**/dist"],
-  "references": [{ "path": "./tsconfig.node.json" }]
+  "exclude": ["node_modules", "**/dist"]
 }
